Package: bail out of validate_deps early if immutable & no results needed
If the repository is not volatile (i.e. it is immutable), we assume it has good
metadata generated, hence we don't need to verify the ebuild metadata ourselves
for bad values.

But if we need some of the results/side-effects of validate_deps, we can
pass force=True to bypass the early return.

Bug: https://bugs.gentoo.org/528394
Signed-off-by: Sam James <sam@gentoo.org>