Begin modernizing the string usage
These files are updated to use the newer f-string feature (String
Interpolation) that has been added since Python 3.6.

The benefits to using String Interpolation are:
* Avoids incurring bugs from using Old Style formatting.
* Makes it easier to understand how the string is being formatted.
* Using embedded expressions into strings removes the need for string
concatenation and will execute much faster for the same effect.

Also, a few global string constants with Old Style formatting had their
literals inlined to their respective function arguments while
transitioning them be f-strings. When Python searches for variables, it
looks them up in the local namespace first, which is the namespace
specific a function or class method. If not found, then it searches in
the global namespace. Removing the need for global strings will also
improve performance a bit along the way.

The Portage codebase as a whole still needs this kind of work.

Signed-off-by: Kenneth Raplee <kenrap@kennethraplee.com>
Signed-off-by: Sam James <sam@gentoo.org>