Prefer generator expressions over list comprehensions
List comprehensions do improve the performance compared to using for
loops to build the same lists in Python. However, they are eagerly
evaluated and will create a sequence ahead of time.

The problem is, a fair amount of code would only use these lists as
iterators which does not require them to be sequences. Unless there is a
need to know a list's length, access its items, or alter its size, it's
better to use generators which are lazily evaluated and can be iterated
when necessary. Avoiding the eagerness of list comprehensions will
improve the runtime performance.

Signed-off-by: Kenneth Raplee <kenrap@kennethraplee.com>
Signed-off-by: Sam James <sam@gentoo.org>