GitSync.update: default to "--depth 1" (bug 824782 comment 17)
Enforce the use of "--depth" in both GitSync.new and GitSync.update following the same logic.

Each function gets its own designated portage option of "clone-depth" and "sync-depth".
This means that portage option "sync-depth" is not taken into consideration anymore in GitSync.new.

Portage option "clone-depth" and "sync-depth" lead to following behaviour with a value of:
- 0: The use of "--depth <value>" is disabled causing the retrieval of the whole Git history.
- ℕ (positive integer): "--depth <value>" is used.
- Unset portage option: "--depth 1" is used.

Bug: https://bugs.gentoo.org/824782
Signed-off-by: David Sardari <d@duxsco.de>
Closes: https://github.com/gentoo/portage/pull/801
Signed-off-by: Sam James <sam@gentoo.org>