Decimate the _unicode_encode() function
Solving this was no easy feat and the diffs in this commit do not show
the hard effort it took to complete. Only 90% of the modules here were
easily solvable by using the Vim editor's regex substitution while
checking for subsitution errors afterward for each module.

The other modules suffered greatly from circular type errors. Those
errors took the longest amount of time to solve. For example, an error
(out of many) would say that an argument to a function needs to be a
bytes object and not the str value it was. After it was encoded to
bytes, a new error says that the same function was looking for a str
instead... And the cycle repeats itself.

I was jumping back and forth to the other modules that came up having
the same problem. During that process, I came up with a tactic using git
to troubleshoot this refactoring process by using a divide-and-conquer
strategy.

Here is how I did it:

1) When every single _unicode_encode() call is ported to the
   str.encode() method throughtout the codebase, revert back the
   _unicode_encode() definition from the root module. This will be a
   temporary place holder for isolating the problem(s).

2) Revert the modules that were affected by the circular errors but keep
   the modules that were successful at passing the test intact. Run the
   tests again to be sure everything is still passing. Once the tests
   have passed, at this point in time the existing modifications are
   consider to be in a "working state." Create a commit to save those
   modules which will help keep them safe and prevent any overlapping
   modifications by you or external factors from git or other programs.
   By doing this, later changes can be reverted back easily when
   necessary.

3) Select a module at random, gradually port each function back to
   str.encode(), and run the tests again to isolate the good ones and
   narrow it down to ones that are the root cause of the circular errors
   If the change causes a circular error, revert the change and work on
   another function or go to another module until the root causes are
   completely isolated. If a test ran passes after a modification, make
   a "fixup" commit and use the interactive git rebase feature to select
   it for fixing up to the last commit since the current changes are
   known to be in a working states from the tests. This makes it easier
   to keep track of successful changes and revert back unsuccessful ones.

4) Keep doing number 3 over and over again, until its been narrowed down
   completely. At this point it's now easy to spot where the problems
   are since these functions are a dead end. Create a local coercion
   closure based on _unicode_encode() to make it clear which function(s)
   are the real problems. After that is completed, remove the
   _unicode_encode() definition from the root module and remove its
   imports in the affected modules.

This process could take almost a whole day to solve. The goal is to
remove those local coercions by completely type annotating the whole
Portage codebase and use a static type-checking analyzer that does a
deep analysis of the types inside the codebase which uses type inference
to make it much easier to discover what the types are from function
arguments and even variables from their local namespaces. As of this
writing the best tools for this are Microsoft's pyright and Google's
pytype. The former currently having better support.

Signed-off-by: Kenneth Raplee <kenrap@kennethraplee.com>