Replace stack management code in ebegin/eend with a counter
At the start of each phase, the count is set to 0.

If the count is greater than 0 when ebegin is called, a warning is
output. The count is incremented.

If the count is equal to 0 when eend is called, a warning is output.
Otherwise, the count is decremented by 1.

At the end of each phase, if the count is greater than 0, a warning is
output.

Signed-off-by: Mike Gilbert <floppym@gentoo.org>