Allow nested ebegin calls using a count variable
At the start of each phase, the count is set to 0.

The count is incremented each time ebegin is called.

If the count is equal to 0 when eend is called, a warning is output.
Otherwise, the count is decremented by 1.

At the end of each phase, if the count is greater than 0, a warning is
output.

Signed-off-by: Mike Gilbert <floppym@gentoo.org>
Closes: https://github.com/gentoo/portage/pull/854
Signed-off-by: Sam James <sam@gentoo.org>