ebuild.5: drop --disable-dependency-tracking in "example" econf
As discussed in #gentoo-qa, it's easier for now to drop this for
consistency rather than add all the extra args (which would be wrong
by itself without rephrasing all the text below to say e.g. "in EAPI ...,
econf doesn't add ...").

So, go for the easiest fix for now so that things make sense, and
we can revisit later.

Signed-off-by: Sam James <sam@gentoo.org>