estrip: avoid spurious NEEDED warning when no ELF installed (< EAPI 7)
If no ELF files are installed, then we're obviously
not going to have a NEEDED file.

The reason this didn't get hit sooner is because
apparently very few of us are actually testing
EAPI 6 actively - which figures.

We only take this path for < EAPI 7 (non-dostrip)
which means we're actually missing the estrip
optimisation in bug #749624 for newer EAPIs!

(Will be fixed in a followup).

Bug: https://bugs.gentoo.org/749624
Bug: https://bugs.gentoo.org/862606
Fixes: bb88e766897f5e7e0b0a10c48cf99a04edb73a40
Signed-off-by: Sam James <sam@gentoo.org>