install-qa-check.d/60pkgconfig: use ver_test to compare versions
Motivied by Matthew's question if the version comparision could be
improved so that, e.g., 25.0 and 25.0.0 are treated as equal.

Note that using ver_test requires that we check the arguments passed
to ver_test conform to PMS version strings. If this is not the case,
then we fall back to the previous behavior and perform a string
comparision.

Thanks-to: Matthew Smith <matthew@gentoo.org>
Signed-off-by: Florian Schmaus <flow@gentoo.org>