80multilib-strict: use read to split QA_MULTILIB_PATHS
Rather than employ the combination of noglob and unquoted expansions to
convert QA_MULTILIB_PATHS to an array, have read handle it instead. This
addresses an instance of SC2086 which, owing to a recent commit in which
I localised IFS, no longer constitutes a legitimate bug. Still, I
elected to employ read because I do not trust other portage contributors
not to repeat the same mistake in the future. Consider it a case of
leading by example, if you will.

Further, check whether QA_MULTILIB_PATHS is an array by examining its
attributes. This requires bash-4.4 at the minimum. I discussed the
matter with Sam James prior and he consented to the use of its features.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>