ecompress: honour the exit status value of guess_suffix()
Presently, the 'PORTAGE_COMPRESS_SUFFIX' variable is assigned by way of
a command substitution in which the guess_suffix() function is called.
However, the variable is simultaneously exported, duly defeating the
error check. The issue is that the value of the '?' parameter will be
the exit status value of the export builtin, which will never fail.
Address this issue by exporting the variable as a separate step.

Further, ensure that die() throws a meaningful exception in the event
that the call to guest_suffix() does fail.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>