man/emerge.1: add clarification to -k and -K
Both -k and -K imply --with-bdeps=n.  While this is already noted
under the documentation for --with-bdeps, a brief reminder under
the documentation for -k and -K is useful.  A user interested in
binary packages might look under -k and -K first and miss the
default behavior with respect to build time dependencies.

Bug: https://bugs.gentoo.org/863422
Closes: https://github.com/gentoo/portage/pull/886
Signed-off-by: Anthony G. Basile <blueness@gentoo.org>
Signed-off-by: Zac Medico <zmedico@gentoo.org>