ecompress: refactor some of the code out into functions
This commit performs some minor refactoring upon the ecompress utility,
the exact nature of which is described herewith.

Relocate the code that attends to the --ignore and --queue options into
two new functions, respectively named do_ignore() and do_queue().

Refrain from interspersing function declarations with code that isn't
function-scoped. Instead, ensure that the function declarations precede
all other commands, with the sole exception of the command that sources
the "isolated-functions.sh" library.

It should be noted that the code contained by the new functions is
over-indented so as to prevent the patch from being incomprehensible.
The indentation shall be adjusted by the next commit.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>