ecompress: ensure IFS= is in effect for read invocations
Where read is given at least one name to assign to, word splitting
mechanics apply. Given the default value of IFS, records will
effectively be trimmed of any surrounding <blank> characters. It is
categorically incorrect to permit this in the course of processing
pathnames. Address the issue by ensuring that the value of IFS is the
null string on each occasion that read is invoked.

Signed-off-for: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>