ecompress: improve the guess_suffix() function
This commit makes some changes to the guest_suffix() function, which are
described herewith.

Localise its variables. Also, localise the '-' parameter so that the
enablement of the "errexit" shell option is scoped to the function.

Don't make use of brace expansion to print a range of numbers. Doing so
requires for the entire range be expanded as a series of words
beforehand. Even for a narrow range of 0..1000, there is a measurable
impact on memory usage. Instead, use a C-style for loop.

Dispense with the clumsy combination of command substitution, echo
invocation and pathname expansion. Instead, iterate over compress?*,
assigning the first existing match to the 'suffix' variable.

Prior to returning, don't print the suffix unless one was discovered,
and do so using printf rather than echo. Further, ensure that the return
value is something other than 0 in the case that no suffix is found.
Note that the caller is now able to invoke die if something goes wrong.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>