ecompress: use the -files0-from option of GNU find
GNU findutils 4.9.0 introduces a new -files0-from option. With this, it is
easier to defend against unfortunate situations such as that depicted below.

$ path="-L"
$ find "$path" -delete

Now there is the option of writing such code as follows.

$ path="-L"
$ printf '%s\0' "$path" | find -files0-from - -delete

That is, by reading a list of NUL-terminated starting points from the
standard input, it becomes impossible for any of them to be erroneously
treated as an option or predicate.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>