isolated-functions.sh: use a conditional expression to parse MAKEOPTS
Not only is -r a non-standard sed(1) option (unlike -E), but the
conditional expression of bash supports extended regular expression
matching. Use it and avoid three subshells into the bargain.

Where the external utilities are executed, neither create a subshell nor
redirect STDERR more than once.

Use the printf builtin instead of echo. Quote the expansion of 'jobs' so
as to eliminate an SC2068 warning.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>