bin/ebuild.sh: Compare BASHOPTS rather than the output of the shopt builtin
Both the __qa_source() and __qa_call() functions capture the output of the shopt
builtin twice, for the purpose of comparing and warning in the case that
something being sourced or executed alters the global shell options without
later restoring them to their prior state.

The output of shopt is verbose in nature and it is somewhat aggravating to see
its output routinely appearing in bash execution traces. As of bash-4.1, a
BASHOPTS variable is available, which contains a colon-separated list of enabled
options, rendering it terse in nature, while still being suitable for the QA
test in question. So, let's use it and avoid incurring two subshells into the
bargain.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>