ebuild.sh: tidy up the processing of QA_INTERCEPTORS slightly
Presently, the "ebuild.sh" unit iterates over the words comprising the
value of the 'QA_INTERCEPTORS' variable while composing and declaring an
intercepting function for each. This commit applies several minor
improvements, which are described herewith.

Refrain from specifying the -f option to the type builtin. It is not
required because type -P will never match against function names.

Specify the "--" operand to the type builtin, signifying end-of-options.

Quote the expansion of the 'BIN' variable to quell an SC2086 warning.

Check the return value of the type builtin to quell an SC2181 warning.

Finally, suppress an SC2068 warning for the unquoted expansion of the
'QA_INTERCEPTORS' variable, for it is safe to perform word splitting for
its potential values.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>