vartree: use ewarn if merging despite collisions
Use ewarn, not eerror, if merge is going ahead despite collisions.

Collisions aren't always fatal - say if FEATURES="protect-owned unmerge-orphans"
and the collision is over an orphaned file.

This changes the colors used once we discover that all relevant files
are orphaned to be warning rather than an error, to make it less scary
and distinguish from other situations.

Closes: https://github.com/gentoo/portage/pull/952
Signed-off-by: Sam James <sam@gentoo.org>