bin/etc-update: Apply patsub_replacement defences
Per bug #881383, string replacing forms of parameter expansion must take care
to quote - or appropriately escape - any nested parameter expansions, assuming
that their values are intended to be taken literally (as is almost invariably
the case). This has long been the case, but the introduction of the new
patsub_replacement option in bash >=5.2 has brought the issue to the fore.

This commit addresses two instances in which the etc-update script could
unintentionally induce patsub replacement. There are many other quality issues
that affect this script but this is enough to address the aforementioned bug.

Bug: https://bugs.gentoo.org/881383
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>