sync: git: only perform shallow updates if the repository is shallow
With the new default of shallow cloning/updating git repositories,
added with f2207e41792d ("GitSync.update: default to "--depth 1" (bug
824782 comment 17)"), existing *non-shallow* repositories would become
shallow after a "emerge --sync".

This adds a check to see if the repository is already shallow, and
only in this case performs a shallow clone.

Thanks to Douglas Freed for pointing out that this should consider
whether or not the user explicitly set sync-depth.

Thanks-to: Douglas Freed <dwfreed@mtu.edu>