sync: git: allow truncating history if repository is marked volatile
See 54a08c9f71dc37851fdb0364576581ca19df6580 for an explanation
of the context here.

If the repository is marked as volatile, it's assumed that
Portage has complete control of it, and for Portage's purposes,
there's no use in having extended history (in fact, if anything,
it's deterimental, as it makes a sync more likely to fail due
to the large disk space requirements over time).

Bug: https://bugs.gentoo.org/824782
Bug: https://bugs.gentoo.org/887025
Signed-off-by: Sam James <sam@gentoo.org>