Simplify signal handling in entry point scripts
These scripts have some boiler-plate code to install a stub handler for
SIGINT, SIGTERM, and SIGPIPE to prevent errors from being displayed to
the user.

For SIGINT, Python installs a handler that raises KeyboardInterrupt. We
reset the handler to SIG_DFL to avoid this.

For SIGTERM, we don't actually need to do anything: Python doesn't
handle SIGTERM, and the process will be terminated silently.

For SIGPIPE, Python ignores this by default, and a BrokenPipeError is
raised if we attempt to write to a broken pipe. We reset the handler
to SIG_DFL to avoid this.

Bug: https://bugs.gentoo.org/887817
Signed-off-by: Mike Gilbert <floppym@gentoo.org>