Rework signal handling in entry scripts
Introduce a new exception SignalInterrupt which inherits from
KeyboardInterrupt and adds a 'signum' member. When a signal is received,
raise SignalInterrupt.

At the end of the script, catch KeyboardInterrupt, and look for the
signum member. Reset the signal handler to SIG_DFL, and re-raise the
signal to kill the process.

This ensures that the invoking shell sees that we got killed by a signal
instead of calling exit.

Bug: https://bugs.gentoo.org/887817
Signed-off-by: Mike Gilbert <floppym@gentoo.org>