bin/quickpkg: loosen default umask
This default was originally changed in bug #182428 to prevent config
files with secrets being leaked into binary packages that were
readable by everyone, but given the default is to not include configs,
it is safe to use a less restrictive umask by default. Also, warn the
user if there is a possibility that secrets end up in a world-readable
binpkg.

Bug: https://bugs.gentoo.org/182428
Bug: https://bugs.gentoo.org/890823
Signed-off-by: John Helmert III <ajak@gentoo.org>