SCHEDULING_POLICY: switch to scheduler policy IDs rather than aliases.
The os.SCHED_* aliases are not reliable enough, the mainline Python lacks
deadline, and the pypy lacks them all together. The IDs are not going to
change, at most new would arrive, so it is safe to use it as is.

By extension, the deadline policy was also added, since those originally
were not present in aliases.

Closes: https://bugs.gentoo.org/867031
Signed-off-by: KARBOWSKI Piotr <slashbeast@gentoo.org>