emaint: Changed default action for emaint target sync to 'auto'
When using 'sync' module of emaint the default option is '--check' when
none is provided which spits out an ERROR about '--check' not being an option for module 'sync':
```
terminal ~ # emaint sync

ERROR: module 'sync' does not have option '--check'

sync module options:
-A, --allrepos (sync module only): -A, --allrepos  Sync all repos that have a sync-url defined
-a, --auto    (sync module only): -a, --auto  Sync auto-sync enabled repos only
-r, --repo    (sync module only): -r, --repo  Sync the specified repo
```

My idea was to change the default action for the 'sync' module to '--auto'. To me '--auto'
seems to be the most useful default option for the 'sync' module since I would imagine
that is the most common use case and presumably most users typing 'emaint sync' would
be looking for this outcome.

Closes: https://github.com/gentoo/portage/pull/981
Signed-off-by: Noel Nyback <noel@nyback.com>
Signed-off-by: Sam James <sam@gentoo.org>