fowners, fperms: Fix handling of relative pathnames
PMS section 12.3.9 specifies that fowners and fperms "[take] paths
relative to the image directory."

Currently, these helpers interpret any pathnames not beginning with a
slash as relative to the current working directory (after printing a
QA warning). This contradicts the spec.

It is also inconsistent with other helpers like insinto or dostrip,
which expand their pathnames relative to ${ED}.

Bug: https://bugs.gentoo.org/905223
Closes: https://github.com/gentoo/portage/pull/1027
Signed-off-by: Ulrich Müller <ulm@gentoo.org>