emerge-webrsync: tidy up 'webrsync-gpg' case
webrsync-gpg is actually two distinct cases:
1. A user has FEATURES="webrsync-gpg" in make.conf and is calling
   'emerge-webrsync'. This is deprecated.

2. A user has 'sync-type=webrsync', is using emaint/emerge to sync,
   and Portage is shelling out to emerge-webrsync. This is what
   users are encouraged to do, but it currently uses the legacy
   webrsync-gpg path.

For the benefit of case #2 and to clarify things, don't mention
FEATURES="webrsync-gpg" if the user hasn't set it - this is to avoid
users *starting to set it* because it sounds like something they may
want.

We also silence the (new) 'gpg legacy path' phrasing for now with
--quiet given this is expected with case #2.

Consolidate some logic while at it.

The next step is, of course, to shift everything to the gemato path.

Bug: https://bugs.gentoo.org/905358
Signed-off-by: Sam James <sam@gentoo.org>