tests: pytest: the last four failing tests with pytest fixed
The issue was: when the full test suite was run, the four tests
would fail. But not when they were run alone. The reason: the
environment where those tests was run on was wrong when the full
suite was run, and not when the individual tests were executed.

It was fixed by adding:

  "PORTAGE_INST_GID": str(os.getuid()),
  "PORTAGE_INST_UID": str(os.getgid()),

to the environments used by the commands exercised in those tests.

Signed-off-by: David Palao <david.palao@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>