emerge-webrsync: create a new temporary dir for legacy gpg verification
It's possible that we can't read /root/.gnupg and we shouldn't
be poking around in there anyway.

However, if the user is setting PORTAGE_GPG_DIR by themselves,
it's their responsibility to handle the directory being in the
right state (e.g. has the right keys imported).

- If PORTAGE_GPG_DIR is unset, make a tmpdir w/ mktemp.
- If we're using that temporary directory we just created, import PORTAGE_GPG_KEY,
as before defaulting to /usr/share/openpgp-keys/gentoo-release.asc.

Bug: https://bugs.gentoo.org/905868
Signed-off-by: Sam James <sam@gentoo.org>