depgraph: Do not allow slotted deps to be satisfied by wrong slots
This may be part of what caused the "perl rebuild bug". The
"priority" of perl, when seen by perl modules, may have "satisfied"
set to an installed perl of a wrong slot. Compounding this factor
with bug #756199 where find_smallest_cycle would select a single
node, in this case because the dependency of perl is satisfied and
the priority then gets ignored, the "cycle" becomes a perl module
alone and gets rebuilt early.

I also updated the test from the previous patch to account for
this change. No other tests seems affected.

For a larger scale test, I reproduced this initially with a stage3
chroot from a Jan 1 2022 stage3 snapshot, and testing in an
equivalent dockerfile would work too:

  FROM gentoo/stage3:amd64-openrc-20220101
  RUN emerge-webrsync
  COPY . /portage

Before this patch (USE flags omitted):

  # cd /portage && bin/emerge -puDN @world 2>&1 | grep -i perl
  [ebuild     U  ] app-admin/perl-cleaner-2.30-r1 [2.30]
  [ebuild  rR    ] virtual/perl-File-Temp-0.231.100
  [ebuild  rR    ] dev-perl/Locale-gettext-1.70.0-r1
  [ebuild  rR    ] dev-perl/MIME-Charset-1.12.2-r1
  [ebuild  rR    ] dev-perl/Module-Build-0.423.100
  [ebuild  rR    ] dev-perl/Text-CharWidth-0.40.0-r2
  [ebuild     U  ] dev-lang/perl-5.36.0-r2 [5.34.0-r3]
  [ebuild  N     ] virtual/perl-CPAN-2.330.0
  [ebuild     U  ] virtual/perl-ExtUtils-MakeMaker-7.640.0 [7.620.0]
  [ebuild     U  ] virtual/perl-File-Spec-3.840.0 [3.800.0]
  [...]

After this patch:

  # cd /portage && bin/emerge -puDN @world 2>&1 | grep -i perl
  [ebuild     U  ] app-admin/perl-cleaner-2.30-r1 [2.30]
  [ebuild     U  ] dev-lang/perl-5.36.0-r2:0/5.36 [5.34.0-r3:0/5.34]
  [ebuild  N     ] virtual/perl-CPAN-2.330.0
  [ebuild     U  ] virtual/perl-ExtUtils-MakeMaker-7.640.0 [7.620.0]
  [ebuild     U  ] virtual/perl-Data-Dumper-2.184.0 [2.179.0]
  [ebuild     U  ] virtual/perl-File-Spec-3.840.0 [3.800.0]
  [ebuild     U  ] virtual/perl-Test-Harness-3.440.0-r1 [3.430.0]
  [ebuild  rR    ] dev-perl/Pod-Parser-1.630.0-r1
  [ebuild  rR    ] dev-perl/Text-CharWidth-0.40.0-r2
  [ebuild  rR    ] dev-perl/Text-WrapI18N-0.60.0-r2
  [...]

Bug: https://bugs.gentoo.org/463976
Bug: https://bugs.gentoo.org/592880
Bug: https://bugs.gentoo.org/596664
Bug: https://bugs.gentoo.org/631490
Bug: https://bugs.gentoo.org/764365
Bug: https://bugs.gentoo.org/793992
Signed-off-by: YiFei Zhu <zhuyifei1999@gmail.com>